Add afj command to analyze jmptbl from the shell ##anall
$ r2 Payload/Undecimus.app/Undecimus
[0x10007e3ac]> s 0x1000081d8;af;s 0x1000082b0;afj 0x1000085fc 4
