+++ libr/anal/Makefile
@@ -35,7 +35,7 @@ OBJLIBS+=anal_ex.o switch.o state.o cycles.o
 OBJLIBS+=esil_sources.o esil_interrupt.o
 OBJLIBS+=esil_stats.o esil_trace.o flirt.o labels.o
 OBJLIBS+=esil2reil.o pin.o session.o vtable.o rtti.o
-OBJLIBS+=rtti_msvc.o rtti_itanium.o
+OBJLIBS+=rtti_msvc.o rtti_itanium.o jmptbl.o
 ASMOBJS+=$(LTOP)/asm/arch/xtensa/gnu/xtensa-modules.o
 ASMOBJS+=$(LTOP)/asm/arch/xtensa/gnu/xtensa-isa.o
 ASMOBJS+=$(LTOP)/asm/arch/xtensa/gnu/elf32-xtensa.o
