+++ libr/core/cmd_anal.c
@@ -285,6 +285,7 @@ static const char *help_msg_af[] = {
 	"afr", " ([name]) ([addr])", "analyze functions recursively",
 	"af+", " addr name [type] [diff]", "hand craft a function (requires afb+)",
 	"af-", " [addr]", "clean all function analysis data (or function at addr)",
+	"afj", " [tableaddr] [count]", "analyze function jmptableble",
 	"afa", "", "analyze function arguments in a call (afal honors dbg.funcarg)",
 	"afb+", " fcnA bbA sz [j] [f] ([t]( [d]))", "add bb to function @ fcnaddr",
 	"afb", "[?] [addr]", "List basic blocks of given function",
@@ -2548,6 +2549,21 @@ static int cmd_anal_fcn(RCore *core, const char *input) {
 				: core->offset;
 			r_anal_fcn_del_locs (core->anal, addr);
 			r_anal_fcn_del (core->anal, addr);
+		}
+		break;
+	case 'j': // "afj"
+		{
+			RAnalFunction *fcn = r_anal_get_fcn_in (core->anal, core->offset, 0);
+			if (fcn) {
+				char *args = strdup (input + 2);
+				RList *argv = r_str_split_list (args, " ");
+				ut64 table = r_num_math (core->num, r_list_get_n (argv, 0));
+				ut64 elements = r_num_math (core->num, r_list_get_n (argv, 1));
+				r_anal_jmptbl (core->anal, fcn, core->offset, table, elements, UT64_MAX);
+				run_pending_anal (core);
+			} else {
+				eprintf ("No function defined here\n");
+			}
 		}
 		break;
 	case 'a': // "afa"
